+++ crates/uv-distribution-types/src/dist_error.rs
@@ -1,3 +1,8 @@
+use crate::{DistRef, Edge, Name, Node, Resolution, ResolvedDist};
+use petgraph::prelude::EdgeRef;
+use petgraph::Direction;
+use rustc_hash::FxHashSet;
+use std::collections::VecDeque;
 use std::fmt::{Display, Formatter};
 use uv_normalize::{ExtraName, GroupName, PackageName};
 use uv_pep440::Version;
@@ -37,6 +42,70 @@ impl FromIterator<DerivationStep> for DerivationChain {
 }
 
 impl DerivationChain {
+    /// Compute a [`DerivationChain`] from a resolution graph.
+    ///
+    /// This is used to construct a derivation chain upon install failure in the `uv pip` context,
+    /// where we don't have a lockfile describing the resolution.
+    pub fn from_resolution(
+        resolution: &Resolution,
+        target: DistRef<'_>,
+    ) -> Option<DerivationChain> {
+        // Find the target distribution in the resolution graph.
+        let target = resolution.graph().node_indices().find(|node| {
+            let Node::Dist {
+                dist: ResolvedDist::Installable { dist, .. },
+                ..
+            } = &resolution.graph()[*node]
+            else {
+                return false;
+            };
+            target == dist.as_ref()
+        })?;
+
+        // Perform a BFS to find the shortest path to the root.
+        let mut queue = VecDeque::new();
+        queue.push_back((target, None, None, Vec::new()));
+
+        // TODO(charlie): Consider respecting markers here.
+        let mut seen = FxHashSet::default();
+        while let Some((node, extra, group, mut path)) = queue.pop_front() {
+            if !seen.insert(node) {
+                continue;
+            }
+            match &resolution.graph()[node] {
+                Node::Root => {
+                    path.reverse();
+                    path.pop();
+                    return Some(DerivationChain::from_iter(path));
+                }
+                Node::Dist { dist, .. } => {
+                    for edge in resolution.graph().edges_directed(node, Direction::Incoming) {
+                        let mut path = path.clone();
+                        path.push(DerivationStep::new(
+                            dist.name().clone(),
+                            extra.clone(),
+                            group.clone(),
+                            dist.version().clone(),
+                            Ranges::empty(),
+                        ));
+                        let target = edge.source();
+                        let extra = match edge.weight() {
+                            Edge::Optional(extra, ..) => Some(extra.clone()),
+                            _ => None,
+                        };
+                        let group = match edge.weight() {
+                            Edge::Dev(group, ..) => Some(group.clone()),
+                            _ => None,
+                        };
+                        queue.push_back((target, extra, group, path));
+                    }
+                }
+            }
+        }
+
+        None
+    }
+
     /// Returns the length of the derivation chain.
     pub fn len(&self) -> usize {
         self.0.len()
