+++ crates/uv-requirements/src/lib.rs
@@ -5,7 +5,7 @@ pub use crate::sources::*;
 pub use crate::specification::*;
 pub use crate::unnamed::*;
 
-use uv_distribution_types::{DerivationChain, Dist, DistErrorKind, GitSourceDist, SourceDist};
+use uv_distribution_types::{Dist, DistErrorKind, GitSourceDist, SourceDist};
 use uv_git::GitUrl;
 use uv_pypi_types::{Requirement, RequirementSource};
 
@@ -20,12 +20,7 @@ pub mod upgrade;
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     #[error("{0} `{1}`")]
-    Dist(
-        DistErrorKind,
-        Box<Dist>,
-        DerivationChain,
-        #[source] uv_distribution::Error,
-    ),
+    Dist(DistErrorKind, Box<Dist>, #[source] uv_distribution::Error),
 
     #[error(transparent)]
     Distribution(#[from] uv_distribution::Error),
@@ -41,24 +36,16 @@ impl Error {
     /// Create an [`Error`] from a distribution error.
     pub(crate) fn from_dist(dist: Dist, cause: uv_distribution::Error) -> Self {
         match dist {
-            Dist::Built(dist) => Self::Dist(
-                DistErrorKind::Download,
-                Box::new(Dist::Built(dist)),
-                DerivationChain::default(),
-                cause,
-            ),
+            Dist::Built(dist) => {
+                Self::Dist(DistErrorKind::Download, Box::new(Dist::Built(dist)), cause)
+            }
             Dist::Source(dist) => {
                 let kind = if dist.is_local() {
                     DistErrorKind::Build
                 } else {
                     DistErrorKind::DownloadAndBuild
                 };
-                Self::Dist(
-                    kind,
-                    Box::new(Dist::Source(dist)),
-                    DerivationChain::default(),
-                    cause,
-                )
+                Self::Dist(kind, Box::new(Dist::Source(dist)), cause)
             }
         }
     }
