+++ crates/uv-resolver/src/error.rs
@@ -106,11 +106,7 @@ pub enum ResolveError {
     ),
 
     #[error("Failed to read metadata from installed package `{0}`")]
-    ReadInstalled(
-        Box<InstalledDist>,
-        DerivationChain,
-        #[source] InstalledDistError,
-    ),
+    ReadInstalled(Box<InstalledDist>, #[source] InstalledDistError),
 
     #[error(transparent)]
     NoSolution(#[from] NoSolutionError),
